requirements.txt
django==1.10.6
djangorestframework==3.6.3
django-filter==1.0.1
markdown==2.6.8
celery==4.0.2
redis==2.10.5
shortuuid==0.5.0
djangorestframework-jwt==1.10.0
PyJWT==1.4.2
django-redis==4.7.0
alidayu==1.0.2
colorlog==2.10.0
emails==0.5.13
jinja2==2.9.5
django-cors-headers==2.0.2
django-annoying==0.10.3
Pillow==4.0.0
xinge-push==1.1.8.3
django-mptt
beautifulsoup4==4.6.0
Epubzilla==0.1.1